+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix alertmanager secretToken in imported alertmanager code.
+
 ## [1.27.1] - 2021-03-25
 
 ### Fixed